Pass actor_id into the scheduler
